requirements/prod.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile requirements/prod.in
#
aiohttp==3.8.3
    # via gradio
aiosignal==1.2.0
    # via aiohttp
analytics-python==1.4.0
    # via gradio
anyio==3.6.1
    # via
    #   httpcore
    #   starlette
async-timeout==4.0.2
    # via aiohttp
asynctest==0.13.0
    # via aiohttp
attrs==22.1.0
    # via aiohttp
backoff==1.10.0
    # via analytics-python
bcrypt==4.0.0
    # via paramiko
certifi==2022.9.24
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via uvicorn
colorama==0.4.5
    # via click
cryptography==38.0.1
    # via paramiko
cycler==0.11.0
    # via matplotlib
fastapi==0.85.0
    # via gradio
ffmpy==0.3.0
    # via gradio
fonttools==4.37.4
    # via matplotlib
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2022.8.2
    # via gradio
gradio==3.4
    # via -r requirements/prod.in
h11==0.12.0
    # via
    #   gradio
    #   httpcore
    #   uvicorn
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via gradio
idna==3.4
    # via
    #   anyio
    #   requests
    #   rfc3986
    #   yarl
importlib-metadata==5.0.0
    # via click
jinja2==3.1.2
    # via gradio
joblib==1.2.0
    # via scikit-learn
kiwisolver==1.4.4
    # via matplotlib
linkify-it-py==1.0.3
    # via markdown-it-py
markdown-it-py[linkify,plugins]==2.1.0
    # via
    #   gradio
    #   mdit-py-plugins
markupsafe==2.1.1
    # via jinja2
matplotlib==3.5.3
    # via gradio
mdit-py-plugins==0.3.1
    # via markdown-it-py
mdurl==0.1.2
    # via markdown-it-py
monotonic==1.6
    # via analytics-python
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
numpy==1.21.6
    # via
    #   -r requirements/prod.in
    #   gradio
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
orjson==3.8.0
    # via gradio
packaging==21.3
    # via matplotlib
pandas==1.3.5
    # via gradio
paramiko==2.11.0
    # via gradio
pillow==9.2.0
    # via
    #   -r requirements/prod.in
    #   gradio
    #   matplotlib
pycparser==2.21
    # via cffi
pycryptodome==3.15.0
    # via gradio
pydantic==1.10.2
    # via
    #   fastapi
    #   gradio
pydub==0.25.1
    # via gradio
pynacl==1.5.0
    # via paramiko
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   analytics-python
    #   matplotlib
    #   pandas
python-multipart==0.0.5
    # via gradio
pytz==2022.4
    # via pandas
pyyaml==6.0
    # via gradio
requests==2.28.1
    # via
    #   analytics-python
    #   gradio
rfc3986[idna2008]==1.5.0
    # via httpx
scikit-learn==1.0.2
    # via sklearn
scipy==1.7.3
    # via
    #   -r requirements/prod.in
    #   scikit-learn
shapely==1.8.4
    # via -r requirements/prod.in
six==1.16.0
    # via
    #   analytics-python
    #   paramiko
    #   python-dateutil
    #   python-multipart
sklearn==0.0
    # via -r requirements/prod.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.20.4
    # via fastapi
threadpoolctl==3.1.0
    # via scikit-learn
torch==1.12.1
    # via -r requirements/prod.in
typing-extensions==4.3.0
    # via
    #   aiohttp
    #   anyio
    #   async-timeout
    #   importlib-metadata
    #   kiwisolver
    #   markdown-it-py
    #   pydantic
    #   starlette
    #   torch
    #   uvicorn
    #   yarl
uc-micro-py==1.0.1
    # via linkify-it-py
urllib3==1.26.12
    # via requests
uvicorn==0.18.3
    # via gradio
websockets==10.3
    # via gradio
yarl==1.8.1
    # via aiohttp
zipp==3.8.1
    # via importlib-metadata
